Move App into its own directory
